refactor(web-universe-components): move slice-email filter to exchange
