+++ packages/manager/modules/web-universe-components/src/string/index.js
@@ -1,13 +1,9 @@
 import angular from 'angular';
 
 import wucSliceContentFilter from './SliceContent';
-import wucSliceEmailFilter from './SliceEmail';
 
 const moduleName = 'wucString';
 
-angular
-  .module(moduleName, [])
-  .filter('wucSliceContent', wucSliceContentFilter)
-  .filter('wucSliceEmail', wucSliceEmailFilter);
+angular.module(moduleName, []).filter('wucSliceContent', wucSliceContentFilter);
 
 export default moduleName;
